Update sync_to_build_id.py
--force will no longer overwrite cobalt_src and the script
no longer requires .git suffix to be consistent (there seem
to be inconsistencies in Cobalt versions).

Also made script python 3 compatible

+++ cobalt/build/sync_to_build_id.py
@@ -12,7 +12,6 @@ from __future__ import print_function
 import argparse
 import json
 import os
-import platform
 import shutil
 import subprocess
 import sys
@@ -66,14 +65,13 @@ def _RunGitCommandReturnExitCode(gitargs, **kwargs):
 
   """
   popen_args = ["git"] + gitargs
-  p = subprocess.Popen(popen_args, stdout=subprocess.PIPE, **kwargs)
-  output = p.stdout.read().splitlines()
-  return p.wait(), output
+  with subprocess.Popen(popen_args, stdout=subprocess.PIPE, **kwargs) as p:
+    output = p.stdout.read().splitlines()
+    return p.wait(), output
 
 
 def main():
-  dev_null_filename = "nul" if platform.system() == "Windows" else "/dev/null"
-  dev_null = open(dev_null_filename, "w")
+  dev_null = open(os.devnull, "w")  # pylint: disable=consider-using-with
   arg_parser = argparse.ArgumentParser(
       description="Syncs to a given Cobalt build id")
   arg_parser.add_argument("buildid", nargs=1)
@@ -97,7 +95,7 @@ def main():
   hashes = json.loads(outer_json["deps"])
   git_root = os.getcwd()
 
-  for relpath, rep_hash in hashes.iteritems():
+  for relpath, rep_hash in hashes.items():
     path = os.path.normpath(os.path.join(git_root, relpath))
     if not os.path.exists(path):
       # No warning in this case, we will attempt to clone the repository in
@@ -120,17 +118,21 @@ def main():
 
     (requested_repo, _) = rep_hash.split("@")
     remote_url = _RunGitCommand(["config", "--get", "remote.origin.url"],
-                                cwd=path)[0] + ".git"
+                                cwd=path)[0]
+    if requested_repo.endswith(".git"):
+      remote_url += ".git"
+
     if remote_url != requested_repo:
-      if args.force:
+      if args.force and path != git_root:
         shutil.rmtree(path)
       else:
         print(("{0} exists but does not point to the requested repo for that "
                "path, {1}. Either replace that directory manually or run this "
-               "script with --force.").format(path, requested_repo))
+               "script with --force. --force will not try to remove the top "
+               "level repository.").format(path, requested_repo))
         return 1
 
-  for relpath, rep_hash in hashes.iteritems():
+  for relpath, rep_hash in hashes.items():
     path = os.path.normpath(os.path.join(git_root, relpath))
 
     # repo_hash has a repo path prefix like this:
